Make remote.ts in its own module.
