+++ sdk-external-modules/remote-ts/package.json
@@ -1,0 +1,31 @@
+{
+    "name": "genezio-remote",
+    "version": "1.0.0",
+    "exports": {
+        ".": {
+            "browser": {
+                "require": "./dist/cjs/index.js",
+                "import": "./dist/esm/index.js"
+            },
+            "react-native": {
+                "require": "./dist/cjs/index.js",
+                "import": "./dist/esm/index.js"
+            },
+            "default": {
+                "require": "./dist/cjs/index.node.js",
+                "import": "./dist/esm/index.node.js"
+            }
+        }
+    },
+    "scripts": {
+        "build:cjs": "tsc -p tsconfig.cjs.json",
+        "build:esm": "tsc -p tsconfig.esm.json",
+        "build": "npm run build:cjs && npm run build:esm && cp package.json ../../build/src/genezio-remote/"
+    },
+    "files": [
+        "dist"
+    ],
+    "description": "",
+    "main": "./dist/cjs/index.js",
+    "module": "./dist/esm/index.js"
+}
