+++ sdk-external-modules/remote-ts/src/index.node.ts
@@ -1,0 +1,3 @@
+import { Remote } from "./remote.node.js";
+
+export { Remote };
