+++ sdk-external-modules/remote-ts/src/index.ts
@@ -1,0 +1,3 @@
+import { Remote } from "./remote.js";
+
+export { Remote };
