+++ sdk-external-modules/remote-ts/tsconfig.cjs.json
@@ -1,0 +1,8 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "target": "ES2019",
+        "module": "CommonJS",
+        "outDir": "../../build/src/genezio-remote/dist/cjs"
+    }
+}
