+++ sdk-external-modules/remote-ts/tsconfig.esm.json
@@ -1,0 +1,9 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "target": "ES2019",
+        "moduleResolution": "nodenext",
+        "module": "nodenext",
+        "outDir": "../../build/src/genezio-remote/dist/esm"
+    }
+}
