+++ src/generateSdk/templates/nodeSdkJs.ts
@@ -1,139 +1,3 @@
-export const nodeSdkJsRemoteBrowser = `/**
-* This is an auto generated code. This code should not be modified since the file can be overwritten
-* if new genezio commands are executed.
-*/
-
-async function makeRequestBrowser(request, url) {
-   const response = await fetch(\`\${url}\`, {
-       method: 'POST',
-       headers: {
-           'Content-Type': 'application/json'
-       },
-       body: JSON.stringify(request),
-   });
-
-   if (!response.ok) {
-       return response.json().then((error) => Promise.reject(error));
-   }
-
-   return response.json();
-}
-
-
-/**
-* The class through which all request to the Genezio backend will be passed.
-*/
-export class Remote {
-   url = undefined;
-   agent = undefined;
-
-   constructor(url) {
-       this.url = url;
-   }
-
-   deserialize(s) {
-       const e = new Error(s.message);
-       e.stack = s.stack
-       e.info = s.info
-       e.code = s.code
-       return e
-   }
-
-   async call(method, ...args) {
-       const requestContent = {"jsonrpc": "2.0", "method": method, "params": args, "id": 3};
-       const response = await makeRequestBrowser(requestContent, this.url);
-
-       if (response.error) {
-           throw this.deserialize(response.error)
-       }
-
-       return response.result;
-   }
-}
-`;
-
-export const nodeSdkJsRemoteNode = `/**
-* This is an auto generated code. This code should not be modified since the file can be overwritten
-* if new genezio commands are executed.
-*/
-
-import * as http from 'http';
-import * as https from 'https';
-
-
-async function makeRequestNode(request, url, agent) {
-   const data = JSON.stringify(request);
-   const hostUrl = new URL(url);
-
-   const options = {
-       hostname: hostUrl.hostname,
-       path: hostUrl.search ? hostUrl.pathname + hostUrl.search : hostUrl.pathname,
-       port: hostUrl.port,
-       method: 'POST',
-       headers: {
-           'Content-Type': 'application/json',
-       },
-       agent: agent,
-   };
-   const client = url.includes('https') ? https : http;
-
-   return new Promise((resolve, reject) => {
-       const req = client.request(options, res => {
-           let body = '';
-
-           res.on('data', d => {
-               body += d
-           });
-           res.on('end', async function() {
-               const response = JSON.parse(body);
-               resolve(response);
-           });
-
-       });
-
-       req.on('error', error => {
-           reject(error);
-       });
-
-       req.write(data);
-       req.end();
-   });
-}
-
-/**
-* The class through which all request to the Genezio backend will be passed.
-*/
-export class Remote {
-   url = undefined;
-   agent = undefined;
-
-   constructor(url) {
-       this.url = url;
-       const client = url.includes("https") ? https : http;
-       this.agent = new client.Agent({ keepAlive: true });
-   }
-
-   deserialize(s) {
-       const e = new Error(s.message);
-       e.stack = s.stack
-       e.info = s.info
-       e.code = s.code
-       return e
-   }
-
-   async call(method, ...args) {
-       const requestContent = {"jsonrpc": "2.0", "method": method, "params": args, "id": 3};
-       const response = await makeRequestNode(requestContent, this.url, this.agent);
-
-       if (response.error) {
-           throw this.deserialize(response.error)
-       }
-
-       return response.result;
-   }
-}
-`;
-
 export const storageJs = `/**
 * This is an auto generated code. This code should not be modified since the file can be overwritten
 * if new genezio commands are executed.
