+++ src/generateSdk/utils/compileSdk.ts
@@ -5,9 +5,10 @@ import ts from "typescript";
 import { Worker } from "worker_threads";
 import { deleteFolder, writeToFile } from "../../utils/file.js";
 import fs from "fs";
+import { default as fsExtra } from "fs-extra";
 import packageManager from "../../packageManagers/packageManager.js";
 import { listFilesWithExtension } from "../../utils/file.js";
-
+import { fileURLToPath } from "url";
 const compilerWorkerScript = `const { parentPort, workerData } = require("worker_threads");
 
 const { compilerOptions, fileNames, typescriptPath } = workerData;
@@ -93,6 +94,15 @@ export async function compileSdk(
     const writePackagePromise = writeToFile(modulePath, "package.json", packageJson, true);
     workers.push(writePackagePromise);
     await Promise.all(workers);
+
+    const __filename = fileURLToPath(import.meta.url);
+    const __dirname = path.dirname(__filename);
+    fsExtra.copySync(
+        path.join(__dirname, "../../genezio-remote"),
+        path.join(genezioSdkPath, "node_modules/genezio-remote"),
+    );
+    fs.writeFileSync(path.join(genezioSdkPath, ".npmignore"), "!node_modules\n");
+
     if (publish === true) {
         console.log("Publishing the SDK...");
         await packageManager.publish(modulePath);
